Move assessment interaction tests
Move the assessment interaction tests to their own module, as the latest
addition made test_interaction.py grow beyond 1000 lines (and fail
pylint).
